requirements.txt
google-generativeai
pyttsx3
streamlit
